refactor: package 분리(util), 코드 목적에 맞는 패키지로 분리
- Constant
- Validator
