+++ src/main/java/racinggame/util/Constants.java
@@ -1,4 +1,4 @@
-package racinggame.domain;
+package racinggame.util;
 
 public class Constants {
